+++ tasks/workflow/worker.go
@@ -216,11 +216,9 @@ func (w *worker) handleExecution(t *taskExecution) {
 		err = w.runScaleIn(ctx, t)
 	case tasks.TaskTypeCustomWorkflow:
 		err = w.runCustomWorkflow(ctx, t, wfName)
-
 	case tasks.TaskTypeAction:
 		err = w.runAction(ctx, t)
-
-	case tasks.TaskTypeQuery, tasks.TaskTypeCustomCommand:
+	case tasks.TaskTypeQuery, tasks.TaskTypeCustomCommand, tasks.TaskTypeForcePurge:
 		// Those kind of task will manage monitoring of taskFailure differently
 		err = w.runOneExecutionTask(ctx, t)
 	default:
@@ -269,6 +267,8 @@ func (w *worker) runOneExecutionTask(ctx context.Context, t *taskExecution) erro
 		err = w.runQuery(ctx, t)
 	case tasks.TaskTypeCustomCommand:
 		err = w.runCustomCommand(ctx, t)
+	case tasks.TaskTypeForcePurge:
+		err = w.runPurge(ctx, t)
 	default:
 		err = errors.Errorf("Unknown TaskType %d (%s) for TaskExecution with id %q", t.taskType, t.taskType.String(), t.taskID)
 	}
@@ -282,9 +282,11 @@ func (w *worker) runCustomCommand(ctx context.Context, t *taskExecution) error {
 	if err != nil {
 		return errors.Wrap(err, "failed to retrieve custom command name")
 	}
-	interfaceNameKv, _, err := kv.Get(path.Join(consulutil.TasksPrefix, t.taskID, "interfaceName"), nil)
+
+	var interfaceName string
+	interfaceName, err = tasks.GetTaskData(kv, t.taskID, "interfaceName")
 	if err != nil {
-		return errors.Wrap(err, "failed to retrieve custom command interface name")
+		return errors.Wrap(err, "failed to retrieve custom interface name")
 	}
 	nodes, err := tasks.GetTaskRelatedNodes(kv, t.taskID)
 	if err != nil {
@@ -294,10 +296,6 @@ func (w *worker) runCustomCommand(ctx context.Context, t *taskExecution) error {
 		return errors.Wrapf(err, "expecting custom command to be related to \"1\" node while it is actually related to \"%d\" nodes", len(nodes))
 	}
 	nodeName := nodes[0]
-	interfaceName := "custom"
-	if interfaceNameKv != nil && len(interfaceNameKv.Value) != 0 {
-		interfaceName = string(interfaceNameKv.Value)
-	}
 	nodeType, err := deployments.GetNodeType(w.consulClient.KV(), t.targetID, nodeName)
 	if err != nil {
 		return err
@@ -571,29 +569,25 @@ func (w *worker) runUndeploy(ctx context.Context, t *taskExecution) error {
 	if status != deployments.UNDEPLOYED {
 		deployments.SetDeploymentStatus(ctx, w.consulClient.KV(), t.targetID, deployments.UNDEPLOYMENT_IN_PROGRESS)
 		t.finalFunction = func() error {
+			defer func() {
+				// in all cases, if purge has been requested, run it at the end
+				if t.taskType == tasks.TaskTypePurge {
+					err := w.runPurge(ctx, t)
+					if err != nil {
+						log.Printf("%+v", err)
+					}
+				}
+			}()
 			_, err := updateTaskStatusAccordingToWorkflowStatus(ctx, t.cc.KV(), t.targetID, t.taskID, "uninstall")
 			if err != nil {
 				return err
 			}
 			// Set it to undeployed anyway
-			err = deployments.SetDeploymentStatus(ctx, w.consulClient.KV(), t.targetID, deployments.UNDEPLOYED)
-			if err != nil {
-				return err
-			}
-			if t.taskType == tasks.TaskTypePurge {
-				err = w.runPurge(ctx, t)
-				if err != nil {
-					return err
-				}
-			}
-			return nil
+			return deployments.SetDeploymentStatus(ctx, w.consulClient.KV(), t.targetID, deployments.UNDEPLOYED)
 		}
 		return w.runWorkflowStep(ctx, t, "uninstall", true)
 	} else if t.taskType == tasks.TaskTypePurge {
-		err = w.runPurge(ctx, t)
-		if err != nil {
-			return err
-		}
+		return w.runPurge(ctx, t)
 	}
 	return nil
 }
@@ -645,10 +639,45 @@ func (w *worker) runPurge(ctx context.Context, t *taskExecution) error {
 }
 
 func (w *worker) runScaleOut(ctx context.Context, t *taskExecution) error {
-	err := deployments.SetDeploymentStatus(ctx, w.consulClient.KV(), t.targetID, deployments.SCALING_IN_PROGRESS)
+	status, err := deployments.GetDeploymentStatus(w.consulClient.KV(), t.targetID)
 	if err != nil {
 		return err
 	}
+	if status == deployments.DEPLOYED {
+		nodeName, err := tasks.GetTaskData(t.cc.KV(), t.taskID, "nodeName")
+		if err != nil {
+			return errors.Wrap(err, "failed to retrieve scale out node name")
+		}
+
+		var instancesDelta int
+		instancesDeltaStr, err := tasks.GetTaskData(t.cc.KV(), t.taskID, "instancesDelta")
+		if err != nil {
+			return errors.Wrap(err, "failed to retrieve scale out instancesDelta")
+		}
+		if instancesDelta, err = strconv.Atoi(instancesDeltaStr); err != nil {
+			return errors.Wrap(err, "failed to convert instancesDelta into int")
+		}
+
+		// Create and store related node instances for scaling
+		instancesByNodes, err := deployments.CreateNewNodeStackInstances(t.cc.KV(), t.targetID, nodeName, instancesDelta)
+		if err != nil {
+			return errors.Wrap(err, "failed to create new nodes instances in topology")
+		}
+		data := make(map[string]string)
+		for scalableNode, nodeInstances := range instancesByNodes {
+			data[path.Join("nodes", scalableNode)] = nodeInstances
+		}
+		err = tasks.SetTaskDataList(t.cc.KV(), t.taskID, data)
+		if err != nil {
+			return errors.Wrap(err, "failed to set task data with nodes for scaling out")
+		}
+
+		err = deployments.SetDeploymentStatus(ctx, w.consulClient.KV(), t.targetID, deployments.SCALING_IN_PROGRESS)
+		if err != nil {
+			return err
+		}
+	}
+
 	t.finalFunction = w.makeWorkflowFinalFunction(ctx, t.cc.KV(), t.targetID, t.taskID, "install", deployments.DEPLOYED, deployments.DEPLOYMENT_FAILED)
 	return w.runWorkflowStep(ctx, t, "install", false)
 }